Extract csv file collection function
